+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+## [Unreleased]
+#### Added
+- Transfer, withdraw and issuance forms disabling if balance is 0
+
+#### Fixed
+- Bug with max value of amount input field for deposit
 
 ## [Unreleased]
 #### Added